+++ src/interceptor_post_entity/index.py
@@ -1,5 +1,5 @@
 from src.interceptor_post_entity.create_error_response import create_error_response
 from src.interceptor_post_entity.extract_data_from_event import extract_data_from_event
 from src.interceptor_post_entity.perform import perform
-from src.interceptor_post_entity.authenticate import authenticate
-from src.utils.authenticate import login
+from src.services.auth.authenticate import authenticate
+from src.services.auth.login import login