+++ src/nonspecific/index.py
@@ -1,6 +1,7 @@
 from src.nonspecific.create_error_response import create_error_response
 from src.nonspecific.extract_data_from_event import extract_data_from_event
 from src.nonspecific.perform import perform
-from src.utils.authenticate import authenticate, login
+from src.services.auth.authenticate import authenticate
+from src.services.auth.login import login
 
-# If you use this hook and get a token from BioT BE you can use the normal auth function here (from util folder)
+# If you use this hook and get a token from BioT BE you can use the normal auth function here (from services/auth folder)